requirements.txt
# requirements.txt

# GUI library for creating the graphical user interface.
tkinter==8.6

# Theme library for enhancing the appearance of Tkinter widgets.
ttkbootstrap==0.1.3

# Library for dealing with time zones.
pytz==2021.3

# Library for serial communication.
pyserial==3.5

# MySQL database connector.
pymysql==1.0.2

# Library for creating and managing SSH tunnels.
sshtunnel==0.4.0

# Provides access to socket operations.
socket==0.1.1

# Module for working with dates and times.
datetime==4.3
